Modyfied code
