+++ src/test/resources/testng_suites/AdminService.xml
@@ -1,18 +1,15 @@
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - OrangeHRM" parallel="methods">
+<suite verbose="1" name="Orange HRM tests - AdminService" parallel="methods">
 
     <parameter name="suiteOwner" value="qpsdemo"/>
-    <parameter name="jenkinsJobName" value="Orange HRM web test"/>
+    <parameter name="jenkinsJobName" value="Orange HRM Admin service web test"/>
     <parameter name="jenkinsJobType" value="web"/>
 
     <parameter name="jenkinsEnvironments" value="DEMO"/>
-    <test name="Orange_HRM web test">
+    <test name="Orange_HRM Admin service web test">
         <classes>
-            <class name="com.solvd.project.carina.demo.OrangeHRMTest">
+            <class name="com.solvd.project.carina.demo.AdminServiceTest">
                 <methods>
-                    <include name="logInTest"/>
-                    <include name="resetPasswordTest"/>
-                    <include name="changePasswordTest"/>
                     <include name="addUserTest"/>
                     <include name="systemUserSearchTest"/>
                 </methods>
