+++ src/test/resources/testng_suites/Authentication.xml
@@ -1,0 +1,21 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite verbose="1" name="Orange HRM tests - Authentication Test" parallel="methods">
+
+    <parameter name="suiteOwner" value="qpsdemo"/>
+    <parameter name="jenkinsJobName" value="Orange HRM authentication web test"/>
+    <parameter name="jenkinsJobType" value="web"/>
+
+    <parameter name="jenkinsEnvironments" value="DEMO"/>
+    <test name="Orange_HRM authentication web test">
+        <classes>
+            <class name="com.solvd.project.carina.demo.AuthenticationTest">
+                <methods>
+                    <include name="logInTest"/>
+                    <include name="resetPasswordTest"/>
+                    <include name="changePasswordTest"/>
+                </methods>
+            </class>
+        </classes>
+    </test>
+
+</suite>
